Update: move hooks to separate file
